Updated readme. Added asserts. Moved app to src folder.
