+++ package.json
@@ -18,10 +18,12 @@
   "homepage": "https://github.com/monstercat/s3-zipper#readme",
   "dependencies": {
     "archiver": "^0.16.0",
+    "assert": "^1.3.0",
     "aws-sdk": "^2.2.10",
     "body-parser": "^1.14.1",
     "debug": "^2.2.0",
     "express": "^4.13.3",
+    "node-uuid": "^1.4.7",
     "request": "^2.65.0"
   }
 }
